+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   An option to return also the empty found tokens was also added via `Tokenizer.KeepEmptyTokens(true)`.
   _This should fix the parsing of whitespace charactes around view query column names when no quotes are used ([#3616](https://github.com/pocketbase/pocketbase/discussions/3616#discussioncomment-7398564))._
 
+- Fixed the `:excerpt(max, withEllipsis?)` `field` query param modifier to properly add space to the generated text fragment after block tags.
+
 
 ## v0.19.0